+++ src/components/App/App.tsx
@@ -1,4 +1,4 @@
-import MainPage from '../../pages/MainPage/MainPage';
+import MainPage from '../../pages/main-page/main-page';
 
 type AppProps = {
   placesCount: number;
