+++ src/pages/login-page/login-page.tsx
@@ -1,4 +1,4 @@
-import Header from '../../components/Header/Header';
+import Header from '../../components/header/header';
 
 function LoginPage (): JSX.Element{
   return(
