+++ src/pages/main-page/tabs/tabs.tsx
